dashboard/backend/requirements.txt
# FastAPI Backend Requirements
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# CORS
fastapi-cors==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
